+++ TraderBot.EmailListener/Program.cs
@@ -3,6 +3,7 @@ using TraderBot.EmailListener.Commands;
 using TraderBot.EmailListener.HostedServices;
 using TraderBot.EmailListener.Infrastructure;
 using TraderBot.OrderControllerProto;
+using TraderBot.RavenDb;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -12,13 +13,10 @@ builder.Services.AddTransient<OpenPositionHandler>();
 builder.Services.AddTransient<ClosePositionHandler>();
 builder.Services.AddTransient<CollectProcessCommand>();
 builder.Services.AddSingleton<EmailChannel>();
-builder.Services.AddSingleton<MailBoxStore>();
 builder.Services.AddHostedService<SchedulerHostedService>();
 builder.Services.AddHostedService<EmailListenerJob>();
 
-builder.Services.AddControllers();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
+await builder.Services.AddAndConfigureRavenDbAsync(builder.Configuration);
 
 builder.Services.AddGrpcClient<OrderControllerGrpc.OrderControllerGrpcClient>((services, options) =>
 {
@@ -29,16 +27,5 @@ builder.Services.AddGrpcClient<OrderControllerGrpc.OrderControllerGrpcClient>((s
 var app = builder.Build();
 
 // Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI(options =>
-    {
-        options.SwaggerEndpoint("/swagger/v1/swagger.json", "EmailListener");
-        options.RoutePrefix = string.Empty;
-    });
-}
-
-app.MapControllers();
 
 app.Run();