+++ trader-bot-view/src/lib/OrdersTable.svelte
@@ -18,6 +18,7 @@
             getOrdersResponse = await client.getOrders(null, null);
         } catch (err) {
             health = false;
+            return;
         }
         orders = getOrdersResponse.response.orders.map((order) => ({
             ...order,
@@ -36,11 +37,11 @@
 
 {#if health}
     <div class="update-indicator">
-        Actual at <a title="Reload" href="javascript:void(0)" on:click={load}>{lastUpdatedAt}</a>
+        Actual at <a title="Reload" href={null} on:click={load}>{lastUpdatedAt}</a>
     </div>
 {:else}
     <div class="error">Communication error</div>
-    <a href="javascript:void(0)" on:click={load}>Try again</a>
+    <a href={null} on:click={load}>Try again</a>
 {/if}
 <div class="orders">
     {#each orders as order}